pyproject.toml
[tool.poetry]
name = "repl_nix_webdevelopment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.3.2"
mysql-connector-python = "^8.0.33"
Flask-Login = "^0.6.2"
Flask-SQLAlchemy = "^3.0.5"
Flask-WTF = "^1.1.1"
WTForms = "^3.0.1"
Flask-Bcrypt = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
